requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-bootstrap4==2.2.0
django-filter==2.3.0
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
